feat: adding a docs sub folder with all the routes
