+++ src/routes/(docs)/docs/zoomcontrol/+page.svelte
